Fixed issue with returning a single project
- Fixed the issue where you could return an array of projects, but not a
single one. This was caused by the "query" entry in apiProject having
the value of true for "isArray". I have added another entry called
"queryOne" to allow only one object to be returned.
- Updated dbProject & dbRelease to use "queryOne" and the projects
controller to use "findOne" instead of "query" and "find" respectively
when returning a single project / release.
- Updated Create Release and View All Releases to display the name of
the project in the heading.