Updated Release data retrieval
- The data is now only added to the scope when it is returned to the
client in View All Releases and View Release.
- The page is now only displayed in View All Releases and View Release
once the release data is added to the scope.
- Updated the releases.js controller to return the appropriate error
codes in getRelease() and getAllReleases(), and to also return the
relevant project data at the same time to reduce the number of client
requests.
- Updated the public View / View All Releases controllers to redirect to
the appropriate error page if an error occurs in the request.