Request tidy up - Project, Feature, User Story, Acceptance Test
Tidied up Project, Feature, User Story, and Acceptance Test to reduce
the number of client requests for data.
- Removed "count" API routes and associated methods, and created an
equivalent in the model. These will be implicitly called by the "get"
routes on the server side and added to the data returned as appropriate.
- Reduced the initial number of calls to the server in each client
controller to a maximum of one.
- Added "exec()" to the end of each mongoose promises, as this is the
recommended approach. It is also required to create a promise of the
aggregate() function.