Created 'View System Test'
- app.js: Updated route titles to have every word starting with a
capital letter, to match the current h1 tags.
- ctrlCreateSystemTest.js, view-all-system-tests.jade, routes.js:
Corrected URLs.
- view-system-test.jade, ctrlViewSystemTest.js, scripts.jade: Created
new view and controller for View System Test, based on the Feature
section.
- system-tests.js: createSystemTest(): The supplied system test code is
now cleared, as the client should not be able to provide this.
- SystemTest.js: Created getNextCode() to determine the code for a new
system test, and corrected createSystemTest() to correctly obtain this
code before creating the object in the database.