Fixes to Add Test Step
- ctrlViewSystemTest: After creating a new test step, the code is now
placed in the scope so that it can be accessed when adding the step to
the system test. When the modal is displayed, the form and search
results are cleared if a search result was previously selected.
- view-system-test: The Create button will not be enabled until the user
has made at least one search for a test step.
- SystemTest & TestStep: Corrected the data type for 'code' so that it
is a Number instead of a String (see UserStory and AcceptanceTest).