Improved selection of creating a test step
- ctrlViewSystemTest: clearTestStepResults: Created a helper method to
clear all test step search results. This will be called when opening up
the modal or when the search criteria changes.
- view-system-test:
-- Added an option to the search results called 'Create new test step'
which is the only way to explicitly create a new step. This should help
to reduce duplicates.
-- The 'Add' button will only appear when an existing search result is
selected from the dropdown. It now also appears as 'Update' if an
existing test step is being updated.
-- The 'Create' button will only appear when the 'Create new test step'
option is selected from the dropdown, or there are no matching search
results.
-- The 'No results found' message will no longer appear before a search
is performed after the search criteria changes.
- site.styl: Added a new 'i' rule to style text in italics.