Started implementing test step table arguments
- ctrlViewSystemTest:
-- isTableArgumentName: Created helper method to check if an argument
name string is for a table (i.e. if it starts and ends with the pipe
symbol).
-- getArgumentNameColumns: Extracts the table header column names from
an argument name string.
-- getArgumentValueRows: Placeholder method which will extract the table
row and column values from an argument value string.
-- updateTestStepsInScope: Updated to add additional values to the
testSteps array in the scope to allow table arguments to be displayed in
the view.
-- clearTestStepSearch: Fixed the call to the 'clearTestStepResults'
method in the scope.
- view-system-test: Added sections to display the table header and table
rows for table arguments.
- site.styl: Added a new style for test step argument table headers.