Implemented test step table arguments
- ctrlViewSystemTest:
-- getArgumentValueFromString: Added a new parameter to allow you to
specify whether a placeholder should be returned if the user has not yet
entered a value.
-- getColumns: Renamed from getArgumentNameColumns.
-- getRows: Renamed from getArgumentValueRows and implemented to extract
the row and column data from a string.
-- updateTestStepsInScope:
--- The argument name is no longer added as its own element to the page.
Instead, it is added as the 'field' attribute to the argument value
element. This will make it easier to figure out what argument the value
on the page relates to instead of trying to get the value of the
previous element in the section (which is fragile).
--- The table argument value is now split up into rows and columns
(using getRows) as 'valueSplit' so that it can be displayed correctly in
the view.
-- submitEdit: Now only used for submitting system test updates (title
and description).
-- submitEditForTestStep: Extracted logic from submitEdit which was only
relevant for test steps to reduce complexity.
-- submitEditForTestStepTable: Calls submitEditForTestStep and manually
regenerates the tables on the view which can't currently be done
automatically by calling the common updateTestStepsInScope method.
- view-system-test:
-- Corrected some indexes ("$parent.index" to "$parent.$parent.index").
-- Updated calls to the submitEdit function and the new variations for
test steps and tables to use the section 'field' value which will now
contain the argument name.
-- Updated the layout for table arguments to allow editing and display a
placeholder if no value has yet been entered.