Implemented 'Associate Feature with System Test'
- ctrlViewSystemTest:
-- updateScope: Created a helper method to update system test expanded
data returned from the server into the scope.
-- updateSystemTest: Created a helper method to send the system test
data in the scope to the server.
-- addFeature: Created a new method to allow a feature code to be
associated with the system test which then submits the updates to the
server.
- add-feature:
-- Changed the binding of the feature results to $index to allow both
the feature's code and name to be easily accessed instead of just the
code.
-- Added an 'Add & Close' button which also closes the popover after
associating the feature with the system test.
- system-tests:
-- getSystemTest: Now gets and returns the features associated with the
system test.
-- updateSystemTest: Now adds the supplied feature codes to the updated
system test if they are strings.
- test-steps: addTestStep: Now also returns the features associated with
the system test.
- Feature: getAllFeatures: Created a method that allows you to get all
features with the supplied project code whose feature code is in the
supplied array. The results are then sorted by name.
- SystemTest: getSystemTestAndTestStepsExpanded:
-- Renamed to getSystemTestExpanded.
-- Now also returns the features associated with the system test.