Navigation can now be expanded & collapsed
nav:
-- The section will now be collapsed by default, only showing linked
entities on a single line. You can collapse this by clicking the button
on the right to put everything onto its own line. This will also show
the section types where there are no links to existing entities so that
you can add new links or click on the button to view all of the entities
in the project.
-- Removed the entity count in each section as this might be misleading.