Updated navigation to on all pages
- *.js: Added the navigation settings for each page to determine the
current section, whether the navigation should be expanded by default,
or whether the associations can be edited.
- *.jade: Replaced the navigation links at the top of each page with a
link to the "nav.jade" partial page.
- nav.jade:
-- Reversed the logic of the setting "isCollapsed" to be "isExpanded"
instead. It can default to false if the value isn't supplied, which is
the case on most pages.
-- The asterisk button is now always added to the page, and only
displayed if it relates to the current section or the navigation is
expanded.
-- Changed the section icon and text to be links matching the location
of the asterisk button.
- site.styl:
-- Removed old "nav" rules.
-- Corrected the heading font sizes after moving the headings on each
page out of "legend" elements.