Implemented association of user stories with acceptance tests
- ctrlViewAcceptanceTest.js: Added new functions to allow adding /
removing of user stories.
- nav.jade: Corrected the display of user stories so that they appear in
the format: 'Project-Feature-UserStory'
- apiUserStory.js, dbUserStory.js, routes.js: Added new search routes
for User Story.
- add-user-story.jade: Corrected the button colour scheme and the
display of search results to use 'As a, I can, So that' format (copy &
paste errors).
- acceptance-tests.js: User Story codes can now be added to the
acceptance test when updating it.
- user-stories.js: searchForUserStory(): Created a new function to
return matching user stories using the project code and text criteria
(feature code, code, as a, I can, so that).
- AcceptanceTest.js:
-- Added 'userStoryCodes' to the schema.
-- getAcceptanceTestExpanded(): Created a new function to return both
the acceptance test data and any linked user stories.
- UserStory.js:
-- getAllUserStories(): Renamed to 'getAllUserStoriesByFeature' for
consistency.
-- getAllUserStories(): Created a new function to return all user
stories with the supplied project code and feature code that have their
code ID in the supplied array.
-- searchForUserStory(): Created a new function to return any matching
user stories by code, feature code, as a, I can, so that.