Project data can now be updated via import
- ctrlViewAllProjects.js: Fixed an issue where redirecting to another
page did not properly close the modal dialog, leaving a 'fade' on the
next page that was obscuring content.
- projects.js:
-- importProject(): Now updates an existing project with the supplied
data.
-- updateProject():
--- Moved some content to a helper method called sanitiseProject().
--- The project code is now returned in the response, as this is used
when redirecting to the imported project.
-- sanitiseProject(): Created a helper method, used by both
importProject() and updateProject() to sanitise the supplied data to
ensure that users can't supply extraneous fields.
- Projects.js:
-- exportProject(): Fixed an issue where the exported project was in an
array instead of an object.
-- importProject(): Removed function as this logic is moving to the
controller in order to more easily access other controller functions.
-- updateProject(): Now also returns the project code of the updated
project (used when importing a project).