Import SpecFlow steps
- header-system-test.jade: Added a new button to the header to allow
test steps to be imported.
- apiStep.js, dbStep.js, routes.js: Added a new 'import' route for step.
- import-step.jade: Created a new partial page (based on
import-project.jade) to allow SpecFlow steps to be imported.
- ctrlViewSystemTest.js: importStep(): Imports the provided SpecFlow
steps into the database.
- view-system-test.jade: Added the new 'import-step' partial page.
- steps.js: importSpecFlowSteps():
-- The project code is now obtained from the posted data instead of a
parameter.
-- Corrected the logic as ProjectStepDefinitions was being treated as a
single object instead of an array.
-- The start and end of the regular expressions are now removed (^ and
$).
- Project.js: Added a new field to the model, lastStepCode, which tracks
the last code assigned to a Step in the project. This is required for
the atomic assignment of codes to steps when importing multiple items at
the same time.
- Step.js:
-- createOrUpdateStep(): Moved to a helper method for re-use
(createOrUpdate()) and extended to allow importing steps without step
codes.
-- importStep(): Created a new method to allow SpecFlow steps to be
imported without supplying existing step codes (since they won't have
any).
-- getNextCode(): Changed to atomically update the last step code stored
against the project. This allows multiple steps to be imported at the
same time without them being assigned the same code and causing an
error.
-- create(): Updated to use the new getNextCode() function.