SpecFlow tables can now be imported
steps.js: importSpecFlowSteps: The argument name for SpecFlow tables is
now appended to the end of the step pattern in a compatible table
format. Since it's not possible to determine what columns should be in
the table, only 2 columns will be created (Key and Value). This should
be enough for most scenarios.