Improved SpecFlow step import efficiency
ctrlViewSystemTest.js: importStep(): Now extracts and only posts the
Project Step Definitions array from the SpecFlow cache (along with the
project code). The Feature Steps node, which contains all the system
test scenarios, is not required. It could potentially be extremely large
so it would be wasteful to post the contents unnecessarily.