Improved responsiveness for Release and Step
- *.pug: Continued setting Bootstrap column classes to improve
responsiveness in the Release and Step sections.
- add...pug: Moved the search buttons on the Add partial views to be on
the right side of the criteria textbox.
- add-step.pug: Moved the Help button to the footer and corrected the
positioning of some "ng-show" attributes to ensure that there is no
excessive padding being displayed when some buttons are hidden.
- create...pug: Changed some Description labels so that they are
contained within a div with the Bootstrap column classes instead of
applying them directly to the label.
- site.styl: Added a new rule to ensure that buttons displayed in
input-group-addon elements will consume all of the available space to
make them easier to click.